Oct 13th commit

+++ opensis-api/opensis.data/Models/SchoolPeriods.cs
@@ -9,14 +9,14 @@ namespace opensis.data.Models
         public int SchoolId { get; set; }
         public int PeriodId { get; set; }
         public decimal? AcademicYear { get; set; }
-        public decimal? SortOrder { get; set; }
+        public int? SortOrder { get; set; }
         public string Title { get; set; }
         public string ShortName { get; set; }
         public decimal? Length { get; set; }
         public string Block { get; set; }
         public string IgnoreScheduling { get; set; }
         public bool? Attendance { get; set; }
-        public decimal? RolloverId { get; set; }
+        public int? RolloverId { get; set; }
         public TimeSpan? StartTime { get; set; }
         public TimeSpan? EndTime { get; set; }
         public DateTime? LastUpdated { get; set; }
